+++ CHANGES.rst
@@ -47,6 +47,9 @@ Changelog
    Commit items for NEXT VERSION
    ..............................
 
+   - feat(package_resource): add module level function get_package_data
+   - refactor: in conditional expression use bool instead of len
+
 .. scriv-start-here
 
 .. _changes_1-6-0-post2:
@@ -73,7 +76,7 @@ Version 1.6.0.post0 — 2025-04-09
 .. _changes_1-6-0:
 
 Version 1.6.0 — 2025-04-08
---------------------------
+------------------------- 
 
 - feat(context_locals): get_locals_dynamic add support for staticmethod and classmethod
 - feat(context_locals): get_locals_dynamic drops arg func_path